updated phonegap to 1.9.0 finished auto add rule function

+++ assets/www/js/netDroidDB.js
@@ -8,11 +8,13 @@
     function getDb() {
         return window.openDatabase('NetDroid', '1.0', 'NetDroid', 1000000);
     }
+
     var _db;
-    $.netDroidDB = function () { };
+    $.netDroidDB = function () {
+    };
     $.extend($.netDroidDB, {
         db:_db ? _db : (_db = getDb()),
-        createRuleTable:function (createSucc) {
+        createRuleTable:function (createSucc,isDropTableBeforeCreate) {
             var db = this.db;
             db.transaction(function (tx) {
                 var createTableSql = 'create table if not exists NetDroidRules(id integer primary key AUTOINCREMENT,' +
@@ -20,13 +22,14 @@
                     'netmask varchar(100) not null,gateway varchar(100) not null,dns1 varchar(100) not null,' +
                     'dns2 varchar(100) not null,sortOrder integer AUTO_INCREMENT,' +
                     "lastModified datetime default (datetime('now','localtime')),status integer default 1);";
-                tx.executeSql('drop table if exists NetDroidRules;');
+                if(isDropTableBeforeCreate)
+                    tx.executeSql('drop table if exists NetDroidRules;');
                 tx.executeSql(createTableSql);
-                tx.executeSql('create unique index NetDroidRules_idx on NetDroidRules(ssidName);');
-            }, function (err) {
-                alert("Error processing SQL: " + err.code + ' message:' + err.message);
-            }, createSucc);
+                tx.executeSql('create unique index if not exists NetDroidRules_idx on NetDroidRules(ssidName);');
+            }, this.errFun, createSucc);
+        },
+        errFun:function (err) {
+            alert("Error processing SQL code: " + err.code + ', message:' + err.message);
         }
     });
-
-}(jQuery, undefined));
+})(jQuery, undefined);
